# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TiendaPOS API",
        "description": "SaaS POS Multi-País para LATAM. Soporta 9 países (VE, CO, MX, EC, AR, PE, CL, BO, UY) con reglas fiscales, IVA, monedas y métodos de pago locales.",
        "contact": {
            "email": "soporte@tiendapos.com"
        },
        "version": "2.1.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/devoluciones": {
            "get": {
                "tags": [
                    "Ventas"
                ],
                "summary": "Listar devoluciones",
                "operationId": "a7cdb8682bf42b6b51e4cb366a03ad8b",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devoluciones paginadas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Ventas"
                ],
                "summary": "Registrar devolución",
                "description": "Procesa una devolución y emite Nota de Crédito.",
                "operationId": "b8b8dc172f1d124bb3d58144bbd34f60",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "venta_id": {
                                        "type": "integer"
                                    },
                                    "almacen_id": {
                                        "type": "integer"
                                    },
                                    "motivo": {
                                        "type": "string"
                                    },
                                    "tipo_reintegro": {
                                        "type": "string",
                                        "enum": [
                                            "efectivo",
                                            "transferencia",
                                            "nota_credito"
                                        ]
                                    },
                                    "moneda": {
                                        "type": "string"
                                    },
                                    "notas": {
                                        "type": "string"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "item_venta_id": {
                                                    "type": "integer"
                                                },
                                                "cantidad": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Devolución procesada"
                    },
                    "422": {
                        "description": "Error de validación o negocio"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/devoluciones/{id}": {
            "get": {
                "tags": [
                    "Ventas"
                ],
                "summary": "Mostrar devolución",
                "operationId": "cc1b0862e3ea8b91109b1c8e87e618bf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devolución con items y venta"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Sanctum token de autenticación. Obténlo via POST /api/v1/auth/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}